fix: 코드 컨벤션 수정 및 QueryDSL dependency 추가
- Service 객체는 Interface 생성 후 Service명Impl로 만들기
- DTO 명은 [Service명]+Request/Response+DTO
