--- src/main/java/org/boot/growup/source/board/dto/ApiResponse.java
@@ -1,5 +1,0 @@
-package org.boot.growup.source.board.dto;
-
-public class ApiResponse {
-
-}
