+++ src/main/java/org/boot/growup/source/board/persist/entity/Notice.java
@@ -1,18 +1,15 @@
 package org.boot.growup.source.board.persist.entity;
 
 import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.boot.growup.source.board.dto.NoticePostDTO;
+import org.boot.growup.source.board.dto.PostNoticeRequestDTO;
 
 @Entity
 @Getter
@@ -24,7 +21,7 @@ public class Notice {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long notice_id;
+  private Long noticeId;
 
   private String title;
   private String content;
@@ -34,10 +31,10 @@ public class Notice {
 //  private Admin admin;
   private String admin;
 
-  public static Notice from(NoticePostDTO noticePostDTO, String admin) {
+  public static Notice of(PostNoticeRequestDTO postNoticeRequestDTO, String admin) {
     return Notice.builder()
-        .title(noticePostDTO.getTitle())
-        .content(noticePostDTO.getContent())
+        .title(postNoticeRequestDTO.getTitle())
+        .content(postNoticeRequestDTO.getContent())
         .admin("admin")
         .build();
   }
