+++ src/main/java/org/boot/growup/source/board/service/NoticeService.java
@@ -1,7 +1,7 @@
 package org.boot.growup.source.board.service;
 
 import lombok.RequiredArgsConstructor;
-import org.boot.growup.source.board.dto.NoticePostDTO;
+import org.boot.growup.source.board.dto.PostNoticeRequestDTO;
 import org.boot.growup.source.board.persist.NoticeRepository;
 import org.boot.growup.source.board.persist.entity.Notice;
 import org.springframework.stereotype.Service;
@@ -18,11 +18,11 @@ public class NoticeService {
    * 공지사항 작성
    */
   @Transactional
-  public Long postNotice(NoticePostDTO noticePostDTO) {
+  public Long postNotice(PostNoticeRequestDTO postNoticeRequestDTO) {
     // 관리자 확인 코드 추가 필요!
 
-    Notice notice = Notice.from(noticePostDTO, "admin");
-    Long id = noticeRepository.save(notice).getNotice_id();
+    Notice notice = Notice.of(postNoticeRequestDTO, "admin");
+    Long id = noticeRepository.save(notice).getNoticeId();
 
     return id;
   }
